+++ src/character/import.js
@@ -500,6 +500,12 @@ export default class CharacterImport extends FormApplication {
         description: "Use D&D Beyond generic item type images, if available (final fallback)",
         enabled: true,
       },
+      {
+        name: "use-full-description",
+        isChecked: game.settings.get("ddb-importer", "character-update-policy-use-full-description"),
+        description: "For actions use full description and snippets, else use snippets only.",
+        enabled: true,
+      }
     ];
 
     const updateReady = (((this.actorOriginal || {}).flags || {}).ddbimporter || {}).inPlaceUpdateAvailable;